--- src/main/java/umc/spring/ringleader/kakao/KakaoService.java
@@ -1,40 +1,0 @@
-package umc.spring.ringleader.kakao;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.http.*;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
-
-@Slf4j
-@Service
-public class KakaoService {
-    private final String GRANT_TYPE= "authorization_code";
-    private final String CLIENT_ID = "09a75fc77502c385b3e7272508ec1663";
-    private final String REDIRECT_URI= "http://localhost:9100/login/kakao";
-    private final String CLIENT_SECRET= "";
-    private final String TOKEN_URL = "https://kauth.kakao.com/oauth/token";
-
-    public String getAccessTokenJsonData(String accessToken){
-        RestTemplate restTemplate = new RestTemplate();
-
-        // 헤더 설정
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("Authorization", "Bearer " + accessToken);
-        HttpEntity request = new HttpEntity(headers);
-
-
-        // 요청 URI과 헤더를 같이 전송
-        ResponseEntity<String> responseEntity = restTemplate.exchange(
-                "https://kapi.kakao.com/v2/user/me",
-                HttpMethod.POST,
-                request,
-                String.class
-        );
-
-        if (responseEntity.getStatusCode() == HttpStatus.OK) {
-            return responseEntity.getBody();
-        }
-        return "error";
-    }
-}
