+++ src/main/java/umc/spring/ringleader/oauth/naver/NaverOAuthService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 
 import lombok.extern.slf4j.Slf4j;
+import umc.spring.ringleader.oauth.utils.OAuthAttributeUtil;
 
 @Slf4j
 @Service
@@ -23,20 +24,6 @@ public class NaverOAuthService {
 			.block();
 
 		log.debug("[USER INFO FROM NAVER] : {}", result);
-		return getEmailFromAttribute(result);
-	}
-
-
-	private String getEmailFromAttribute(String attribute) {
-		Pattern p = Pattern.compile("([\\w\\-]([\\.\\w])+[\\w]+@([\\w\\-]+\\.)+[A-Za-z]{2,4})");
-		Matcher m = p.matcher(attribute);
-
-		while (m.find()) {
-			if (m.group(1) != null) {
-				break;
-			}
-		}
-
-		return m.group(1);
+		return OAuthAttributeUtil.getEmailFromAttribute(result);
 	}
 }
